+++ HISTORY.rst
@@ -5,6 +5,8 @@ Changelog
 ++++++++++++++++++
 
 * Field-level validation.
+* Add ``fields.Method``.
+* Allow binding of extra data to a serialized object by passing the ``extra`` param when initializing a ``Serializer``.
 
 0.1.0 (2013-11-10)
 ++++++++++++++++++